:truck:rename: styles 폴더 생성, 반응형 파일 생성
