+++ client/src/styles/GlobalVariables.js
